actualizacion

+++ backend-spring-boot/src/main/java/com/sword/trailersdb/repositories/MovieRepository.java
@@ -3,6 +3,6 @@ package com.sword.trailersdb.repositories;
 import com.sword.trailersdb.data.models.MovieModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieRepository extends JpaRepository<MovieModel, Long> {
+public interface MovieRepository extends JpaRepository<MovieModel, Integer> {
         
 }
